+++ evm/src/main/java/com/horizen/evm/interop/EvmParams.java
@@ -1,4 +1,4 @@
-package com.horizen.evm.library;
+package com.horizen.evm.interop;
 
 import com.horizen.evm.utils.Address;
