+++ evm/src/main/java/com/horizen/evm/interop/HandleParams.java
@@ -1,4 +1,6 @@
-package com.horizen.evm.library;
+package com.horizen.evm.interop;
+
+import com.horizen.evm.JsonPointer;
 
 public class HandleParams extends JsonPointer {
     public int handle;
